<fix>[block_storage]: specify storage configuration in install path
Resolves: ZSTAC-71948
