+++ kvmagent/kvmagent/plugins/vm_plugin.py
@@ -66,6 +66,7 @@ from zstacklib.utils import drbd
 from zstacklib.utils.jsonobject import JsonObject
 from zstacklib.utils.linux import is_virtual_machine
 from zstacklib.utils.plugin import TaskManager, TaskResult
+from zstacklib.utils import rbd
 from zstacklib.utils.qga import *
 from zstacklib.utils import jsonobject
 from zstacklib.utils.qmp import get_block_node_name_and_file
@@ -1749,7 +1750,7 @@ class IsoCeph(object):
         for minfo in self.iso.monInfo:
             e(source, 'host', None, {'name': minfo.hostname, 'port': str(minfo.port)})
         if self.iso.fsId:
-            e(source, 'config', None, {'file': linux.get_config_path_from_fs_id(self.iso.fsId)})
+            e(source, 'config', None, {'file': rbd.get_config_path_from_fs_id(self.iso.fsId)})
         e(disk, 'target', None, {'dev': target_dev, 'bus': target_bus_type})
         if bus and unit:
             e(disk, 'address', None, {'type': 'drive', 'bus': bus, 'unit': unit})
@@ -1805,7 +1806,7 @@ class VirtioCeph(object):
         for minfo in self.volume.monInfo:
             e(source, 'host', None, {'name': minfo.hostname, 'port': str(minfo.port)})
         if self.volume.fsId:
-            e(source, 'config', None, {'file': linux.get_config_path_from_fs_id(self.volume.fsId)})
+            e(source, 'config', None, {'file': rbd.get_config_path_from_fs_id(self.volume.fsId)})
         e(disk, 'target', None, {'dev': 'vd%s' % self.dev_letter, 'bus': 'virtio'})
         if self.volume.physicalBlockSize:
             e(disk, 'blockio', None, {'physical_block_size': str(self.volume.physicalBlockSize)})
